# profile/api/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Profile API
  description: API микросервиса профилей пользователей
  version: 1.0.0

servers:
  - url: /api/v1
    description: Profile service

paths:
  /profile/me:
    get:
      summary: Получить свой профиль
      operationId: getMe
      tags: [Profile]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ОК
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Неавторизован
        '404':
          description: Профиль не найден
    put:
      summary: Создать/обновить свой профиль
      operationId: upsertMe
      tags: [Profile]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertMeRequest'
      responses:
        '200':
          description: ОК
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Неавторизован

  /profile/{login}:
    get:
      summary: Получить профиль пользователя по логину
      operationId: getByLogin
      tags: [Profile]
      parameters:
        - in: path
          name: login
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ОК
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          description: Профиль не найден
        '400':
          $ref: '#/components/responses/BadRequest'

  /profile/search:
    get:
      summary: Поиск профиля по логину (точно или похожие)
      operationId: searchByLogin
      tags: [Profile]
      parameters:
        - in: query
          name: login
          required: true
          schema:
            type: string
          description: Логин для поиска
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: ОК
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UpsertMeRequest:
      type: object
      properties:
        display_name:
          type: string
          example: "Ivan Ivanov"
        avatar_url:
          type: string
          example: "https://cdn.example.com/avatars/1.png"
        bio:
          type: string
          example: "Hello!"

    Profile:
      type: object
      properties:
        user_id:
          type: string
          description: UUID пользователя (как строка)
          example: "9b0e6f6d-1fda-4a23-8b5a-8a0c9d2d5c20"
        login:
          type: string
          example: "user_123"
        display_name:
          type: string
        avatar_url:
          type: string
        bio:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    SearchResponse:
      type: object
      properties:
        exact:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Profile'
        matches:
          type: array
          items:
            $ref: '#/components/schemas/Profile'

  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
